Particle fixes + small improvements (#174)
* Particle fixes + small improvements

-New batch renderer for flash particles to support additive blending. Config option to disable shaders when rendering it so it still looks ok with the lighting mod. Fire particle now also disables depth so the flamethrower particles don't fail the depth test.

-Write active bending data to NBT so you don't have to switch then around again on world join

-Floating blocks rotate around the player a bit better, and they will now do a ray trace to help pick the most accurate path when thrown. Also fixed crash when the floating block isn't near the player when the spin becomes active.

-Comment questions that I'm not entirely sure about: Aren't orbit ids unnecessary when they already have two other entity ids to organize them by? Doesn't sendToAllTracking do the same thing as the tracking code in AvatarPlayerData#sendPacket?

* Fix particle collisions

Changed getAliveParticles to get the correct particles field and work with the flash particle renderer. Also caches it so instead of getting a new list every time it's called, it only has to get it once per tick.

* Remove print statement

* Fix access modifier

+++ src/main/java/com/crowsofwar/avatar/entity/EntityShockwave.java
@@ -120,7 +120,7 @@ public class EntityShockwave extends EntityOffensive {
 	}
 
 	public double getParticleSpeed() {
-		return dataManager.get(SYNC_PARTICLE_SPEED);
+		return (double) dataManager.get(SYNC_PARTICLE_SPEED);
 	}
 
 	public void setParticleSpeed(float speed) {
@@ -128,7 +128,7 @@ public class EntityShockwave extends EntityOffensive {
 	}
 
 	public double getRange() {
-		return dataManager.get(SYNC_RANGE);
+		return (double) dataManager.get(SYNC_RANGE);
 	}
 
 	public void setRange(float range) {
@@ -136,7 +136,7 @@ public class EntityShockwave extends EntityOffensive {
 	}
 
 	public double getSpeed() {
-		return dataManager.get(SYNC_SPEED);
+		return (double) dataManager.get(SYNC_SPEED);
 	}
 
 	public void setSpeed(float speed) {
